+++ examples/client-object-list-scan-basic/src/main.cpp
@@ -18,7 +18,7 @@
 #endif
 
 #ifndef APP_VERSION
-#define APP_VERSION "0.17.0"
+#define APP_VERSION "0.18.0"
 #endif
 
 #ifndef MY_USE_DHCP